server/nottinghill/requirements.txt
asgiref==3.3.1
Django==3.1.4
django-cors-headers==3.6.0
django-environ==0.4.5
django-rest-passwordreset==1.2.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.7.1
PyJWT==2.1.0
pytz==2020.5
sqlparse==0.4.1
